Third Stage recreation.gov Scraping
Add ability to filter pandas df for the trip start date continuing
through the number of days we want to stay; accomplished with
simple datetime/timedelta.

Add ability to parse filtered df to check if every relevant date
column contains at least one availability (including if the
availabilities are not neccessarily contiguous by row). Had to
manually loop through df columns for this, but there might be a
better/cleaner way to accomplish it.

This commit should be the core functionality of the
recreation.gov scraper, as we can now go from URL to availability
boolean.

TODO: the next part is going to be a mess of integration. Need to
write something to hit ridb, get all campgrounds, then use the
scraper to parse availability for each all in one go. After that,
need some kind of notification ability.