Fixes for Raspberry Pi + Stability After Testing
This commit has been tested in an actual usage situation that
successfully found a campground availability that popped up due to
a cancelation in July 2021.

It was always my intention to run this tool on a Raspberry Pi, but
due to convenience I did all of the initial development on macOS.
Unfortunately, RPi is still 32bit (for the foreseeable future) and
Chromeium/chrome does not have a valid 32bit browser/webdriver
combo because they stopped updating 32bit versions of the webdriver
all the way back in version 2.8. Instead, on an RPi, users have to
download chromium/selenium as a separate install step, rather than
relying on WebDriverManager to pull a webdriver down dynamically.
For the sake of consistency, I have stopped using WebDriverManager
entirely until this situation is resolved on the RPi end. Some docs
have been added to the README about this in the install step, but
need to be expanded

Other changes for stability include using WebDriverWait to
dynamically wait for page elements to load (or for loading tags to
disappear), which avoids the need for hard-coded wait times in
scrape_availability. References have been added for all of this
in-line as well as in README.

TODO: test this install on macOS to be reasonably sure that the
'stability' improvements aren't solely applicable to
linux/raspbian. There's also a lot of dead code/cleanup things
that need to happen before a potential release is put out.