First Release Candidate
This commit has been confirmed to work on macOS and there should
have been no substantive changes from the last commit that was
tested on linux/raspbian. Because we have had a successful real-
world test (run on raspbian) that was able to find a campground
opening and send a notification email, it's time to consider
tagging a release so that users are able to know this is a stable
place in development.

All files have been cleaned up to conform to python best practices
and linting (there are a few places where linting is ignored due to
practical considerations). All comments and inline docstrings have
been updated to reflect the current codebase. The README should
be sufficient for anybody with minimal technical knowledge to
install and run the project, and should contain enough detail/refs
to enable another developer to start making changes.