Enable Closing the Tutorial Window
recreation.gov seems to always load a tutorial page when we start
up a new chromium-browser/chromedriver instance. Sometimes, this
appears to prevent us from finding the start/end date boxes. Add
ability to close the tutorial box.

TODO: this appears to work when running chromium-browser with an
x11 window -- can both see it working visually and also the script
succeeds. Without the x11 window, it looks like the script fails
but still not quite sure why this is happening.