Confirm Text-via-email Works
Followed some guides to do text-via-email gateway to different
carriers. Tested for AT&T and it seems like it works. Most guides
used asyncio, which isn't a bad idea and kind of dovetails with
issue #16, but unsure if we want to commit to that right now given
that everything is still happening synchronously.

- daemon.py: added send_text_alert function + carrier map
- utils.py: reformat logs to see if we can clean them up a little

TODO:

- decide about asyncio
- dedup sending email and text alert code (lots of overlap)
- code cleanup + docs (a lot more needs to be rewritten to
  accomodate latest changes)