Notification By Text (#18)
Followed some guides to do text-via-email gateway to different
carriers. Tested for AT&T and it seems like it works. Most guides
used asyncio, which isn't a bad idea and kind of dovetails with
issue #16, but we aren't committing to asyncio yet because it's
difficult and everything else still happens synchronously so the
efficiency gains would be minimal.

Main changes:
- add table for carrier-to-email translation to daemon
- dedup'd email and text notifications
- add ability to specify the number of required sites
- improve docs and a bunch of stuff for listing